\name{RegClust-package}
\alias{RegClust}
\docType{package}
\title{
A clustering process via regression.
}

\description{
 A clustering process built upon linear regression analysis (Qin and Self 2006), maximum likelihood, Bayesian inference criteria, and Bootstrapping maximum volume.
}

\details{
\tabular{ll}{
Package: \tab RegClust\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-02-17\cr
License: \tab  Artistic-2.0\cr
}
}

\author{
Weichao Bao, Xin Tong, Meredith Ray, Hongmei Zhang
}

\references{
Qin, Li-Xuan, and Steven G. Self. "The clustering of regression models method with applications in gene expression data." Biometrics 62.2 (2006): 526-533.
}

